Renombrando componentes que ya sólo tienen una versión
