+++ frontend/www/js/omegaup/components/contest/Print.vue
@@ -19,7 +19,7 @@
 <script lang="ts">
 import { Vue, Component, Prop } from 'vue-property-decorator';
 import { types } from '../../api_types';
-import problem_SettingsSummary from '../problem/SettingsSummaryV2.vue';
+import problem_SettingsSummary from '../problem/SettingsSummary.vue';
 import omegaup_Markdown from '../Markdown.vue';
 
 @Component({
