+++ frontend/www/js/omegaup/components/problem/Details.vue
@@ -297,12 +297,12 @@ import * as time from '../../time';
 import * as ui from '../../ui';
 import arena_ClarificationList from '../arena/ClarificationList.vue';
 import arena_EphemeralGrader from '../arena/EphemeralGrader.vue';
-import arena_Runs from '../arena/Runsv2.vue';
+import arena_Runs from '../arena/Runs.vue';
 import arena_RunSubmitPopup from '../arena/RunSubmitPopup.vue';
 import arena_RunDetailsPopup from '../arena/RunDetailsPopup.vue';
 import arena_Solvers from '../arena/Solvers.vue';
 import problem_Feedback from './Feedback.vue';
-import problem_SettingsSummary from './SettingsSummaryV2.vue';
+import problem_SettingsSummary from './SettingsSummary.vue';
 import problem_Solution from './Solution.vue';
 import qualitynomination_DemotionPopup from '../qualitynomination/DemotionPopup.vue';
 import qualitynomination_PromotionPopup from '../qualitynomination/PromotionPopup.vue';
